refactor: DTO 패키지 통합
- 분리되어있던 DTO 패키지 통합
