+++ src/main/java/roomescape/controller/AdminReservationController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.service.ReservationService;
-import roomescape.service.dto.result.ReservationResult;
+import roomescape.dto.ReservationResult;
 
 import java.util.List;
