+++ src/main/java/roomescape/controller/ReservationController.java
@@ -5,13 +5,13 @@ import java.net.URI;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import roomescape.controller.dto.ReservationModifyRequest;
+import roomescape.dto.ReservationModifyRequest;
 import roomescape.service.ReservationService;
-import roomescape.service.dto.command.ReservationCreateCommand;
-import roomescape.service.dto.command.ReservationModifyCommand;
-import roomescape.service.dto.result.AvailableDateResult;
-import roomescape.service.dto.result.ReservationResult;
-import roomescape.service.dto.result.ReservationTimeStatusResult;
+import roomescape.dto.ReservationCreateCommand;
+import roomescape.dto.ReservationModifyCommand;
+import roomescape.dto.AvailableDateResult;
+import roomescape.dto.ReservationResult;
+import roomescape.dto.ReservationTimeStatusResult;
 
 import java.time.LocalDate;
 import java.util.List;
