+++ src/main/java/roomescape/dto/ReservationModifyCommand.java
@@ -1,11 +1,14 @@
-package roomescape.controller.dto;
+package roomescape.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
 import org.springframework.format.annotation.DateTimeFormat;
 
-public record ReservationModifyRequest(
+public record ReservationModifyCommand(
+        @NotNull(message = "RESERVATION_ID_NULL")
+        Long reservationId,
+
         @NotNull(message = "PERSON_NAME_NULL_OR_BLANK")
         @NotBlank(message = "PERSON_NAME_NULL_OR_BLANK")
         String name,
