+++ src/main/java/roomescape/dto/ReservationModifyRequest.java
@@ -1,14 +1,11 @@
-package roomescape.service.dto.command;
+package roomescape.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
 import org.springframework.format.annotation.DateTimeFormat;
 
-public record ReservationModifyCommand(
-        @NotNull(message = "RESERVATION_ID_NULL")
-        Long reservationId,
-
+public record ReservationModifyRequest(
         @NotNull(message = "PERSON_NAME_NULL_OR_BLANK")
         @NotBlank(message = "PERSON_NAME_NULL_OR_BLANK")
         String name,
