+++ src/main/java/roomescape/dto/ReservationTimeCreateCommand.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto.command;
+package roomescape.dto;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalTime;
