+++ src/main/java/roomescape/service/ReservationService.java
@@ -12,13 +12,13 @@ import roomescape.exception.ErrorCode;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 import roomescape.repository.ThemeRepository;
-import roomescape.service.dto.command.ReservationCreateCommand;
-import roomescape.service.dto.command.ReservationModifyCommand;
-import roomescape.service.dto.result.AvailableDateResult;
-import roomescape.service.dto.result.ReservationResult;
-import roomescape.service.dto.result.ReservationTimeResult;
-import roomescape.service.dto.result.ReservationTimeStatusResult;
-import roomescape.service.dto.result.ThemeResult;
+import roomescape.dto.ReservationCreateCommand;
+import roomescape.dto.ReservationModifyCommand;
+import roomescape.dto.AvailableDateResult;
+import roomescape.dto.ReservationResult;
+import roomescape.dto.ReservationTimeResult;
+import roomescape.dto.ReservationTimeStatusResult;
+import roomescape.dto.ThemeResult;
 
 import java.time.LocalDate;
 import java.util.List;
