+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -7,8 +7,8 @@ import roomescape.exception.BusinessException;
 import roomescape.exception.ErrorCode;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
-import roomescape.service.dto.command.ReservationTimeCreateCommand;
-import roomescape.service.dto.result.ReservationTimeResult;
+import roomescape.dto.ReservationTimeCreateCommand;
+import roomescape.dto.ReservationTimeResult;
 
 import java.util.List;
