+++ src/main/java/roomescape/service/ThemeService.java
@@ -7,8 +7,8 @@ import roomescape.exception.BusinessException;
 import roomescape.exception.ErrorCode;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ThemeRepository;
-import roomescape.service.dto.command.ThemeCreateCommand;
-import roomescape.service.dto.result.ThemeResult;
+import roomescape.dto.ThemeCreateCommand;
+import roomescape.dto.ThemeResult;
 
 import java.time.LocalDate;
 import java.util.List;
