+++ src/test/java/roomescape/controller/ThemeControllerTest.java
@@ -7,7 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.jdbc.Sql;
 import roomescape.ClearDbTest;
-import roomescape.service.dto.result.ThemeResult;
+import roomescape.dto.ThemeResult;
 
 import java.util.HashMap;
 import java.util.List;
