+++ src/test/java/roomescape/service/ReservationTimeServiceTest.java
@@ -19,8 +19,8 @@ import roomescape.exception.BusinessException;
 import roomescape.exception.ErrorCode;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
-import roomescape.service.dto.command.ReservationTimeCreateCommand;
-import roomescape.service.dto.result.ReservationTimeResult;
+import roomescape.dto.ReservationTimeCreateCommand;
+import roomescape.dto.ReservationTimeResult;
 
 @ExtendWith(MockitoExtension.class)
 class ReservationTimeServiceTest {
